pyproject.toml
[project]
name = "disks-sync"
version = "0.1.0"
description = "Синхронизация Google Drive → Яндекс Диск"
requires-python = ">=3.10"
dependencies = [
    "yadisk>=3.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
disks-sync = "main:main"

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "ty>=0.0.24",
]
